Refactoring and packages update
