+++ src/domain-info.ts
